:knife: Allow tagging elastic-agents
TL;DR
=====
- Elastic-agents supporting single word tags; adding an ability to add
  these in a runtime configuration'
- Added a `runtime.yml`

Detail
======
Nodes deployed by BOSH are often hard to identify purely by the hostname
that they are registered in Fleet with. Tagging allows us to _at least_
have the understanding of what environment/foundation they belong to (or
add tags that are helpful in some other meaningful way).

The README was updated and an example manifest was also updated to
provide some guidance on how to apply tags when the elastic-agent
enrolls.