:knife: Opinionated default tagging, logging fix
TL;DR
=====
- It's important to have the BOSH deployment &
  instance name in Fleet to know what the node is;
  add this by default to each elastic-agent
- First major release of the tile
- Allow `elastic-agent run` to place its logs in
  `/var/vcap/sys/log/elastic-agent`
- Bump runtime config release dependency
- Correct usage of `-` and `_` for variables that
  are exposed in a runtime tile (not included as
  part of this commit)

Detail
======
We're _now_ including the `bosh-instance` and
`bosh-deployment` tags in the elastic-agent
registration to allow the agents to be easier to
discover in the Fleet interface of Kibana. These
values are parsed from flat files in
`/var/vcap/instance`.

Elastic-agent's `ndjson` logs are now placed in
`/var/vcap/sys/log/elastic-agent` to mimic the
behavior of other products deployed by BOSH and
make it easier for the operator to debug the
process if/when it fails.

```
jumpbox/4e7489bb-8905-41d0-84ae-8ae65402d52a:/var/vcap/sys/log/elastic-agent# ls *ndjson
elastic-agent-20240202-1.ndjson  elastic-agent-20240202.ndjson
```

Hyphens are problematic in tiles because they are
automatically converted to underscores when they
are part of a property path. We've updated our
spec to use underscores exclusively, so we don't
run into odd issues when generating a tile from
the runtime configuration.

**NOTE**: The tile isn't yet included as part of
this repo yet, but it will be when we generate an
appropriate CI job to generate it. This will also
include an automated BOSH release generation job,
to coincide with the Tile release.