🔪 Use native systemd configuration
TL;DR
=====
* Move from `elastic-agent enroll` to
  `elastic-agent install`
* `elastic-agent install` uses systemd, so the
  BOSH release has been updated to also use it
* Update default tag format; `key: value` format
  does not get rendered correctly when the
  `--delay-enroll` flag is used at install time
* Bump release to v2.0.0
* Update example runtime config
* Simplify pre-start/ctl scripts

Detail
======
Elastic-agents can be upgraded when they are
installed with a systemd unit file. This is only
possible, presently, by running `elastic-agent install`.
Previously, we were invoking the `enroll` command
to connect a standalone agent to the Fleet Server,
but this inhibits us from being able to upgrade
the agents via Fleet Server's UI, and thus reduces
the functionality and ease of using Fleet Server.
We have since switched to using the native install
method and allow `monit` to monitor the `systemd`
process via regex ("elastic-agent"). To add
additional arguments to the poorly made unit file,
we now introduce a unit file override that adds
the updated logging path to make use of the
standard BOSH `/var/vcap/sys/log/<package>`
location (`/var/vcap/sys/log/elastic-agent`).

During the process of migrating to the native
install mechanism, we discovered `--delay-enroll`
does not fare well when changing the configuration
path (`--path.config`). Instead, the installation
will revert to the path that the elastic-agent
binary is executed from (in our case, this was in
`/var/vcap/packages/elastic-agent`) and is unable
to find the file in the updated config path. To
workaround this limitation (bug?) we are only
redefining the log path, and we leverage the
defaults for everything else.

This is a breaking change, so the release is being
bumped to v2.0.0. The current release will
remain on a v1.X.X tag _if_ we plan to keep that
approach available. The runtime config example is
updated, but Github actions haven't been
introduced yet to automate that process. This is
presently automated in a private Gitlab, but that
work needs to be ported here.

The `pre-start` & `ctl` scripts have been
simplified. Some superfluous steps were present,
and they were not DRY (reused functions). There is
now a simple set of shared functions that can be
used between the different shell scripts